Added in LDAP authentication for the application.
For now, it can be seen a placeholder as adding out own form wouldn't be too hard.
Also, the management of users is quite different this time, because of the nature of the app. If a user is expired, they will not be returned when queried - but their entry remains in the database. Useful to figure out who has logged in.
Logouts are done as soon as somebody goes to the view/ page.

Signed-off-by: Victory000 <victoria.a.lacroix@gmail.com>